feat: 证书管理-支持证书分配给业务-后端 --story=137438279 (merge request !3404)
Squash merge branch 'cert-assign-to-business' into 'bcc/v1.9.x'
feat: 证书管理-支持证书分配给业务-后端 --story=137438279


TAPD: --story=137438279

+++ cmd/data-service/service/cloud/cert/update.go
@@ -110,7 +110,7 @@ func batchUpdateCertExt[T corecert.Extension](cts *rest.Contexts, svc *certSvc) 
 		for _, item := range *req {
 			updateData := &tablecert.SslCertTable{
 				Name:             item.Name,
-				BkBizID:          int64(item.BkBizID),
+				BkBizID:          item.BkBizID,
 				Domain:           item.Domain,
 				CertType:         item.CertType,
 				CertStatus:       item.CertStatus,
